CBMC: Call mld_randombytes by contract in crypto_sign_keypair
For the crypto_sign_keypair, the wrong function name was used in the
USE_FUNCTION_CONTRACTS. mld_randombytes() is the function that comes with a
contract which internally calls randombytes() which does not have a contract.
In consequence, CBMC would inline mlk_randombytes, which forwards to
randombytes which has no contract attached to it. Being asked to call
randombytes by contract, CBMC would create an empty contract rather than
failing.

In consequence, the CBMC proofs previously did not check the validity of the
output buffer passed to mlk_randombytes.

This commit fixes the issue by changing randombytes to mlk_randombytes in the
CBMC Makefile.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>